refactor: extract out delete instrument
